Update Landed Cost

+++ landed_cost/models/landed_cost.py
@@ -143,6 +143,8 @@ class StockLandedCost(models.Model):
                     
                     
     def compute_landed_cost(self):
+        if len(self.picking_ids)>1:
+            raise ValidationError(_("Transfer count must be single"))
         AdjustementLines = self.env['stock.valuation.adjustment.lines']
         AdjustementLines.search([('cost_id', 'in', self.ids)]).unlink()
         self['cost_lines'].unlink()
@@ -190,10 +192,22 @@ class StockLandedCost(models.Model):
                 value_split = 0.0
                 for valuation in cost.valuation_adjustment_lines:
                     value = 0.0
+                    unit_price = total_cost = invoice_total = 0.0
                     if valuation.cost_line_id and valuation.cost_line_id.id == line.id:
                         if line.split_method == 'by_quantity' and total_qty:
-                            per_unit = (line.price_unit / total_qty)
-                            value = valuation.quantity * per_unit
+                            unit_price_valuation = cost.picking_ids.move_ids.filtered(lambda x:x.product_id==valuation.product_id).stock_valuation_layer_ids
+                            if len(unit_price_valuation)>1 or len(unit_price_valuation)==0:
+                                raise ValidationError(_("Product %s has many valuation layers")%(valuation.product_id.name))
+                            else:
+                                unit_price = unit_price_valuation.unit_cost
+                            total_cost = unit_price * valuation.quantity
+                            invoice_total = sum(cost.picking_ids.move_ids.stock_valuation_layer_ids.mapped('value'))
+                            if not invoice_total:
+                                raise ValidationError(_("There is no Invoice Total to Calculate"))
+                            else:
+                                value = round((line.price_unit/invoice_total)*total_cost,0)
+                            # per_unit = (line.price_unit / total_qty)
+                            # value = valuation.quantity * per_unit
                         elif line.split_method == 'by_weight' and total_weight:
                             per_unit = (line.price_unit / total_weight)
                             value = valuation.weight * per_unit
